requirements.txt
# Core Backend Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangGraph & AI Agents (simplified for now)
langgraph
langchain-core

# HTTP Client
httpx==0.25.2

# Database
psycopg2-binary==2.9.9

# Environment
python-dotenv==1.0.0

# Development
pytest==7.4.3